add support for the CPUID instruction
Exposes a device file to userspace called "cpuid", which contains the formatted
contents of the results of the CPUID instruction. This lists all the features
and extended features supported by this particular processor. This information
can be used directly by the user to know what machine xv6 is running on, or can
also be used in the kernel to load additional drivers or take advantage of
particular cpu features.

Signed-off-by: Joel Nider <joeln@il.ibm.com>